added updates for cross-platform support

+++ Ghostscript.NET/gs/iapi.h.cs
@@ -39,8 +39,8 @@ namespace Ghostscript.NET
     {
         public IntPtr product;
         public IntPtr copyright;
-        public int revision;
-        public int revisiondate;
+        public long revision;
+        public long revisiondate;
     }
 
     #endregion
